backend/requirements.txt
fastapi[standard]==0.112.0
uvicorn==0.32.0
pillow
numpy==1.22.4
pandas==2.2.3
scikit-learn==1.3.0
python-dotenv==1.0.1
opencv-python==4.8.0.74  # Adjust version as needed
matplotlib==3.8.0  # Adjust version as needed
# Specify the version explicitly
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.2.2+cu118
torchvision==0.17.2+cu118
onnx==1.16.0